Split libp2p tests below file size gate
