+++ crates/oasis7_net/src/libp2p_net/tests/discovery_dial_tests.rs
@@ -1,0 +1,50 @@
+use super::signed_discovery_peer_record;
+use crate::libp2p_net::PeerManagerPolicy;
+use libp2p::PeerId;
+use libp2p::identity::Keypair;
+use std::collections::{HashMap, HashSet};
+
+#[test]
+fn process_discovered_peer_record_dials_candidate_peer() {
+    crate::libp2p_net::runtime_support::run_on_libp2p_test_runtime(|| {
+        let mut swarm = crate::libp2p_net::swarm_behaviour::build_swarm(
+            &Keypair::generate_ed25519(),
+            false,
+            true,
+            std::time::Duration::from_secs(30),
+            crate::libp2p_net::wire_bytes::init_shared_wire_byte_counters(),
+        );
+        let peer_key = Keypair::generate_ed25519();
+        let record = signed_discovery_peer_record(
+            &peer_key,
+            vec![
+                crate::dht::PeerDiscoverySource::Dht,
+                crate::dht::PeerDiscoverySource::Rendezvous,
+            ],
+            1,
+        );
+        let peer_id = PeerId::from(peer_key.public());
+        let mut discovered_peer_records = HashMap::new();
+        let mut known_transport_paths = HashMap::new();
+        let mut last_dialed_transport_paths = HashMap::new();
+        let active_transport_paths = HashMap::new();
+        let mut failed_transport_path_labels = HashSet::new();
+
+        crate::libp2p_net::discovery::process_discovered_peer_record(
+            &mut swarm,
+            &mut discovered_peer_records,
+            &mut known_transport_paths,
+            &mut last_dialed_transport_paths,
+            &active_transport_paths,
+            &mut failed_transport_path_labels,
+            None,
+            &PeerManagerPolicy::default(),
+            record.clone(),
+        )
+        .expect("process candidate peer record");
+
+        assert!(discovered_peer_records.contains_key(&peer_id));
+        assert!(known_transport_paths.contains_key(&peer_id));
+        assert!(last_dialed_transport_paths.contains_key(&peer_id));
+    });
+}
